requirements.txt
-i https://pypi.org/simple
awsebcli==3.15.3
blessed==1.16.1
botocore==1.12.253
cached-property==1.5.1
cement==2.8.2
certifi==2019.9.11
chardet==3.0.4
colorama==0.3.9
django-extra-fields==2.0.0
django==2.2.6
djangorestframework==3.10.3
docker-compose==1.23.2
docker-pycreds==0.4.0
docker==3.7.3
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
future==0.16.0
idna==2.7
jmespath==0.9.4
jsonschema==2.6.0
pathspec==0.5.9
pillow==6.2.1
psycopg2-binary==2.8.4
pygments==2.4.2
python-dateutil==2.8.0 ; python_version >= '2.7'
pytz==2019.3
pyyaml==3.13
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.0
termcolor==1.1.0
texttable==0.9.1
urllib3==1.24.3 ; python_version >= '3.4'
wcwidth==0.1.7
websocket-client==0.56.0
